+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - use the templating maven plugin to indicate the version of project
  - worker returns void value
  - create a new class: MessageSenderHelper to send easily a message in as2
+ - supports AES-128, AES-192 and AES-256 algorithms.
 
 ## 1.1 (2015-07-06)
  - add loggers